requirements.txt.txt
pyTelegramBotAPI
google-genai
Pillow